% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censys-package.r
\docType{package}
\name{censys}
\alias{censys}
\alias{censys-package}
\title{Tools to access the Censys API}
\description{
Censys is a search engine that allows computer scientists to ask questions
about the devices and networks that compose the Internet. Driven by I
nternet-wide scanning, Censys lets researchers find specific hosts and
create aggregate reports on how devices, websites, and certificates are
configured and deployed.
}
\details{
The Censys API provides programmatic access to the same data accessible
through web interface (\url{https://www.censys.io/}).

Censys tutorial: \url{https://www.censys.io/tutorial}
}
\author{
Bob Rudis (@hrbrmstr)
}
\references{
\url{https://www.censys.io/about};
            \url{https://www.censys.io/static/censys.pdf}
}

