% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{censys_export_download}
\alias{censys_export_download}
\title{Download export job files to a specified directory}
\usage{
censys_export_download(job_id, path)
}
\arguments{
\item{job_id}{Censys export job id (from calling \code{censys_start_export()})}

\item{path}{Location for downloaded data.}
}
\value{
API call result (invisibly)
}
\description{
Download export job files to a specified directory
}
\examples{
\dontrun{
q <- censys_start_export("
SELECT location.country, count(ip) FROM ipv4.20161206 GROUP BY location.country
")
censys_export_job_status(q$job_id)
censys_export_download(q$job_id, "~/Data")
}
}

