% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlquery.r
\name{censys_get_job_result}
\alias{censys_get_job_result}
\title{Get results of completed Censys SQL query job}
\usage{
censys_get_job_result(job_id, page = 1)
}
\arguments{
\item{job_id}{Censys job id (from calling \code{censys_query()})}

\item{page}{page number of paged results}
}
\description{
The Get Results endpoint allows you to retrieve results of a query after it has
completed successfully.
}
\references{
Censys SQL query syntax: \url{https://censys.io/query};
            API doc: \url{https://www.censys.io/api/v1/docs/search}
}

