% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlquery.r
\name{censys_get_job_status}
\alias{censys_get_job_status}
\title{Get status of a Censys SQL query job}
\usage{
censys_get_job_status(job_id)
}
\arguments{
\item{job_id}{Censys job id (from calling \code{censys_query()})}
}
\value{
API call result (invisibly)
}
\description{
The Get Job Status endpoint allows you to determine whether a job has completed. Once
it has successfully finished, you can then retrieved results with the Get Results
endpoint. Data should be posted as a JSON request document.
}
\references{
Censys SQL query syntax: \url{https://censys.io/query};
            API doc: \url{https://www.censys.io/api/v1/docs/search}
}

