% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlquery.r
\name{censys_series}
\alias{censys_series}
\title{List all series that can be queried from the SQL interface}
\usage{
censys_series()
}
\description{
You must have both \code{CENSYS_API_ID} and \code{CENSYS_API_SECRET} present in the
R environment for the functions in this package to work. It is highly suggested that
you place those in \code{~/.Renviron} at least for interactive work.
}
\examples{
\dontrun{
censys_series()
}
}
\references{
Censys SQL query syntax: \url{https://censys.io/query};
            API doc: \url{https://www.censys.io/api/v1/docs/search}
}

