% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlquery.r
\name{censys_series_details}
\alias{censys_series_details}
\title{Get details about a series, including the list of tables and schema for the series}
\usage{
censys_series_details(series = "ipv4")
}
\arguments{
\item{series}{series name (call \code{censys_series() to see them all}). Defaults to \code{ipv4}.}
}
\description{
You must have both \code{CENSYS_API_ID} and \code{CENSYS_API_SECRET} present in the
R environment for the functions in this package to work. It is highly suggested that
you place those in \code{~/.Renviron} at least for interactive work.
}
\examples{
\dontrun{
censys_series_details("ipv4")
}
}
\references{
Censys SQL query syntax: \url{https://censys.io/query};
            API doc: \url{https://www.censys.io/api/v1/docs/search}
}

