% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlquery.r
\name{censys_query}
\alias{censys_query}
\title{Issue SQL Queries against the Censys API}
\usage{
censys_query(sql)
}
\arguments{
\item{query}{SQL query string}
}
\value{
API call result (invisibly)
}
\description{
The Query API allows executing SQL queries against our daily snapshots and raw data
analogous to the Query web interface. Queries are executed asynchronously. You must
first start a job, then check its status. Once a job has completed, you can view
paginated results using the get results endpoint. Jobs typically require 15-30
seconds to execute; results can be viewed for 24 hours after the job completed.
Ddefinition endpoints are also exposed where you can list the series and view series
details (i.e., list tables and schema).
}
\examples{
\dontrun{
}
}
\references{
Censys SQL query syntax: \url{https://censys.io/query};
            API doc: \url{https://www.censys.io/api/v1/docs/search}
}

